Change init/teardown to the event loop starting/stopping
NagMQ has always initialized at the beginning of the process,
which means that it could potentially initialize before all
of nagios's worker processes start, which means there are
stale ZeroMQ sockets lying around causing problems with binds
later on. This will move the init/teardown to the event loop
starting/stopping, which gives a better guarantee that
Nagios is in a sane state when NagMQ is trying to bind to things.