Reverting/fixing off-by-one error. Adding reconnect_ivl to mqexec
The off-by-one error fixed in the last commit is better fixed by
setting the initial size of the hash table appropriately.

This also adds the reconnect_ivl (reconnect_interval) option to
mqexec, and sets the default reconnect interval to 1 second
instead of 100 ms. This prevents connection storms when keys
aren't yet added to nagios, but clients are trying to connect.