Change event broker logging to nagios log functions
The NagMQ event broker has always logged very sparsely to syslog,
which drops messages and confuses users by having to look in
a different place for NagMQ messages than for Nagios messages.

This commit changes error/warning messages to use the Nagios log,
and the debug messages to go to the nagios debug log.